pyproject.toml
[tool.poetry]
name = "activity-detection"
version = "0.1.0"
description = ""
authors = ["Shourya Sharma <shourya.sharma@complyadvantage.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.41.2"
einops = "^0.8.0"
setuptools = "^70.0.0"
opencv-contrib-python-headless = "^4.10.0.82"
pillow = "^10.3.0"
torch = "^2.3.1"
torchvision = "^0.18.1"
ultralytics = "^8.2.31"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
pyobjc = "^10.3.1"
gtts = "^2.5.1"
opencv-python = "^4.10.0.84"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pytest = "^8.2.2"
pre-commit = "^3.7.1"
pytest-cov = "^5.0.0"

[tool.poetry.scripts]
activity-detection = "activity_detection.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
